requirements-dev.txt
# Development and testing dependencies for Power Pet Door integration

# Core testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-timeout>=2.0.0

# Home Assistant testing
pytest-homeassistant-custom-component

# Snapshot testing (optional)
syrupy>=4.0.0

# Mocking is provided by unittest.mock in Python 3 standard library
